feat: revert breaking change, add tests for wrapNestedPath
