+++ lib/lambda.js
@@ -535,9 +535,29 @@ function getLambdaHandlerInfo (env, modules, logger) {
   }
 }
 
+// wraps a nested property on the module object with the lambda handler
+// @param string field ex. "foo.baz.bar"
+// @param object module
+function wrapNestedPath (agent, field, module) {
+  const parts = field.split('.')
+  let object = module
+  while (parts.length > 0) {
+    const prop = parts.shift()
+    if (parts.length === 0 && object[prop]) {
+      object[prop] = agent.lambda(object[prop])
+    } else {
+      object = object[prop]
+    }
+    if (!object) {
+      break
+    }
+  }
+}
+
 module.exports = {
   isLambdaExecutionEnvironment,
   elasticApmAwsLambda,
   setLambdaTransactionData,
-  getLambdaHandlerInfo
+  getLambdaHandlerInfo,
+  wrapNestedPath
 }
