+++ test/lambda/wrapper.test.js
@@ -3,7 +3,7 @@ const tape = require('tape')
 const path = require('path')
 const Instrumentation = require('../../lib/instrumentation')
 const logging = require('../../lib/logging')
-const { getLambdaHandlerInfo } = require('../../lib/lambda')
+const { getLambdaHandlerInfo, wrapNestPath, wrapNestedPath } = require('../../lib/lambda')
 
 const MODULES = Instrumentation.modules
 tape.test('unit tests for getLambdaHandlerInfo', function (suite) {
@@ -143,4 +143,34 @@ tape.test('integration test', function (t) {
   // did normal patching/wrapping take place
   t.equals(express.static.name, 'wrappedStatic', 'express module was instrumented correctly')
   t.end()
+})
+
+tape.test('wrapper', function (suite) {
+  suite.test('basic case', function(t){
+    // wrapNestedPath (agent, field, module) {
+    const agent = {
+      lambda:function(prop) {
+        t.equals(prop, 'bar', 'expected value passed')
+        t.end()
+      }
+    }
+    const module = {
+      foo:'bar',
+    }
+    wrapNestedPath(agent, 'foo', module)
+  })
+
+  suite.test('property does not exist', function(t){
+    const agent = {
+      lambda:function(prop) {
+        t.fail()
+      }
+    }
+    const module = {
+      foo:'bar',
+    }
+    wrapNestedPath(agent, 'fooo', module)
+    t.end()
+  })
+  suite.end()
 })
